+++ Source/CommonPluginProcessor.h
@@ -97,7 +97,7 @@ public:
     osci::BooleanParameter* muteParameter = nullptr;
 
     std::shared_ptr<osci::Effect> volumeEffect = std::make_shared<osci::SimpleEffect>(
-        [this](int index, osci::Point input, const std::vector<std::atomic<double>>& values, double sampleRate) {
+        [this](int index, osci::Point input, const std::vector<std::atomic<float>>& values, float sampleRate) {
             volume = values[0].load();
             return input;
         }, new osci::EffectParameter(
@@ -109,7 +109,7 @@ public:
     );
 
     std::shared_ptr<osci::Effect> thresholdEffect = std::make_shared<osci::SimpleEffect>(
-        [this](int index, osci::Point input, const std::vector<std::atomic<double>>& values, double sampleRate) {
+        [this](int index, osci::Point input, const std::vector<std::atomic<float>>& values, float sampleRate) {
             threshold = values[0].load();
             return input;
         }, new osci::EffectParameter(
