+++ Source/audio/PolygonBitCrushEffect.h
@@ -5,12 +5,12 @@
 // Inspired by xenontesla122
 class PolygonBitCrushEffect : public osci::EffectApplication {
 public:
-    osci::Point apply(int index, osci::Point input, const std::vector<std::atomic<double>> &values, double sampleRate) override {
+    osci::Point apply(int index, osci::Point input, const std::vector<std::atomic<float>>&values, float sampleRate) override {
         const double pi = juce::MathConstants<double>::pi;
         const double twoPi = juce::MathConstants<double>::twoPi;
-        double effectScale = juce::jlimit(0.0, 1.0, values[0].load());
-        double nSides = juce::jmax(2.0, values[1].load());
-        double stripeSize = juce::jmax(1e-4, values[2].load());
+        double effectScale = juce::jlimit(0.0f, 1.0f, values[0].load());
+        double nSides = juce::jmax(2.0f, values[1].load());
+        double stripeSize = juce::jmax(1e-4f, values[2].load());
         stripeSize = std::pow(0.63 * stripeSize, 1.5); // Change range and bias toward smaller values
         double rotation = values[3] * twoPi;
         double stripePhase = values[4];
