+++ Source/audio/VortexEffect.h
@@ -3,10 +3,10 @@
 
 class VortexEffect : public osci::EffectApplication {
 public:
-    osci::Point apply(int index, osci::Point input, const std::vector<std::atomic<double>> &values, double sampleRate) override {
+    osci::Point apply(int index, osci::Point input, const std::vector<std::atomic<float>>&values, float sampleRate) override {
         // Treat input as complex number and raise to integer power
         // Disallowing non-integer and negative exponents because of the branch cut
-        double effectScale = juce::jlimit(0.0, 1.0, values[0].load());
+        double effectScale = juce::jlimit(0.0f, 1.0f, values[0].load());
         double exponent = juce::jmax(1.0, std::floor(values[1].load() + 0.001));
         double refTheta = values[2].load() * juce::MathConstants<double>::twoPi;
